+++ src/main/java/com/techelevator/Candy.java
@@ -1,8 +1,10 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Candy extends Item {
 	public Candy(String slotIdentifier, String name, double price, String type) {
-		super(slotIdentifier, name, price, type);
+		super(slotIdentifier, name, BigDecimal.valueOf(price), type);
 	}
 @Override
 	public String getMessage() {
